переделал dao убрал лишнее
