+++ hw12-WebServer/src/main/java/ru/otus/hw12/WebServerWithBasicSecurityDemo.java
@@ -5,16 +5,15 @@ import com.google.gson.GsonBuilder;
 import org.eclipse.jetty.security.HashLoginService;
 import org.eclipse.jetty.security.LoginService;
 import org.hibernate.SessionFactory;
-import ru.otus.hw12.dao.DbServiceWebServer;
 import ru.otus.hw12.dao.UserDao;
-import ru.otus.hw12.dao.UserDaoAdapter;
+import ru.otus.hw12.dao.UserDaoWebServerAdapterFromMyHibernate;
 import ru.otus.hw12.helpers.FileSystemHelper;
-import ru.otus.hw12.hibernate.core.model.Address;
-import ru.otus.hw12.hibernate.core.model.Phone;
-import ru.otus.hw12.hibernate.core.model.User;
-import ru.otus.hw12.hibernate.hibernate.HibernateUtils;
-import ru.otus.hw12.hibernate.hibernate.dao.UserDaoHibernate;
-import ru.otus.hw12.hibernate.hibernate.sessionmanager.SessionManagerHibernate;
+import ru.otus.hw12.dao.hibernate.core.model.Address;
+import ru.otus.hw12.dao.hibernate.core.model.Phone;
+import ru.otus.hw12.dao.hibernate.core.model.User;
+import ru.otus.hw12.dao.hibernate.hibernate.HibernateUtils;
+import ru.otus.hw12.dao.hibernate.hibernate.dao.UserDaoHibernate;
+import ru.otus.hw12.dao.hibernate.hibernate.sessionmanager.SessionManagerHibernate;
 import ru.otus.hw12.server.UsersWebServer;
 import ru.otus.hw12.server.UsersWebServerWithBasicSecurity;
 import ru.otus.hw12.services.TemplateProcessor;
@@ -31,8 +30,7 @@ public class WebServerWithBasicSecurityDemo {
                 "hibernate.cfg.xml", User.class, Address.class, Phone.class);
         SessionManagerHibernate sessionManager = new SessionManagerHibernate(sessionFactory);
         UserDaoHibernate userDaoHibernate = new UserDaoHibernate(sessionManager);
-        UserDaoAdapter userDaoAdapter = new UserDaoAdapter(userDaoHibernate);
-        UserDao userDao = new DbServiceWebServer(userDaoAdapter);
+        UserDao userDao = new UserDaoWebServerAdapterFromMyHibernate(userDaoHibernate);
         Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
         TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
