--- hw12-WebServer/src/main/java/ru/otus/hw12/cachehw/HwListener.java
@@ -1,9 +1,0 @@
-package ru.otus.hw12.cachehw;
-
-/**
- * @author sergey
- * created on 14.12.18.
- */
-public interface HwListener<K, V> {
-  void notify(K key, V value, String action);
-}
