--- hw12-WebServer/src/main/java/ru/otus/hw12/dao/UserDaoWebServer.java
@@ -1,12 +1,0 @@
-package ru.otus.hw12.dao;
-
-import ru.otus.hw12.hibernate.core.dao.UserDao;
-import ru.otus.hw12.hibernate.core.model.User;
-
-import java.util.List;
-
-public interface UserDaoWebServer extends UserDao {
-
-    List<User> getAllUsers();
-
-}
