+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/cachehw/HwListener.java
@@ -1,0 +1,9 @@
+package ru.otus.hw12.dao.hibernate.cachehw;
+
+/**
+ * @author sergey
+ * created on 14.12.18.
+ */
+public interface HwListener<K, V> {
+  void notify(K key, V value, String action);
+}
