+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/dao/UserDao.java
@@ -1,0 +1,16 @@
+package ru.otus.hw12.dao.hibernate.core.dao;
+
+
+
+import ru.otus.hw12.dao.hibernate.core.model.User;
+import ru.otus.hw12.dao.hibernate.core.sessionmanager.SessionManager;
+
+import java.util.Optional;
+
+public interface UserDao {
+  Optional<User> findById(long id);
+
+  long saveUser(User user);
+
+  SessionManager getSessionManager();
+}
