+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/dao/UserDaoException.java
@@ -1,0 +1,7 @@
+package ru.otus.hw12.dao.hibernate.core.dao;
+
+public class UserDaoException extends RuntimeException {
+  public UserDaoException(Exception ex) {
+    super(ex);
+  }
+}
