+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/model/Address.java
@@ -1,0 +1,30 @@
+package ru.otus.hw12.dao.hibernate.core.model;
+
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "addresses")
+public class Address {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @Column(name = "id")
+    private long id;
+
+    @Column(name = "address")
+    private String street;
+
+    public Address() {
+    }
+
+    public Address(String street) {
+        this.street = street;
+    }
+
+    @Override
+    public String toString() {
+        return "street='" + street;
+    }
+
+}
