+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/service/DbServiceException.java
@@ -1,0 +1,7 @@
+package ru.otus.hw12.dao.hibernate.core.service;
+
+public class DbServiceException extends RuntimeException {
+  public DbServiceException(Exception e) {
+    super(e);
+  }
+}
