+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/sessionmanager/DatabaseSession.java
@@ -1,0 +1,4 @@
+package ru.otus.hw12.dao.hibernate.core.sessionmanager;
+
+public interface DatabaseSession {
+}
