+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/sessionmanager/SessionManager.java
@@ -1,0 +1,10 @@
+package ru.otus.hw12.dao.hibernate.core.sessionmanager;
+
+public interface SessionManager extends AutoCloseable {
+    void beginSession();
+    void commitSession();
+    void rollbackSession();
+    void close();
+
+    DatabaseSession getCurrentSession();
+}
