+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/core/sessionmanager/SessionManagerException.java
@@ -1,0 +1,12 @@
+package ru.otus.hw12.dao.hibernate.core.sessionmanager;
+
+
+public class SessionManagerException extends RuntimeException {
+  public SessionManagerException(String msg) {
+    super(msg);
+  }
+
+  public SessionManagerException(Exception ex) {
+    super(ex);
+  }
+}
