+++ hw12-WebServer/src/main/java/ru/otus/hw12/dao/hibernate/hibernate/sessionmanager/DatabaseSessionHibernate.java
@@ -1,0 +1,31 @@
+package ru.otus.hw12.dao.hibernate.hibernate.sessionmanager;
+
+import org.hibernate.Session;
+import org.hibernate.Transaction;
+import ru.otus.hw12.dao.hibernate.core.sessionmanager.DatabaseSession;
+
+
+public class DatabaseSessionHibernate implements DatabaseSession {
+  private final Session session;
+  private final Transaction transaction;
+
+  DatabaseSessionHibernate(Session session) {
+    this.session = session;
+    this.transaction = session.beginTransaction();
+  }
+
+  public Session getHibernateSession() {
+    return session;
+  }
+
+  public Transaction getTransaction() {
+    return transaction;
+  }
+
+  public void close() {
+    if (transaction.isActive()) {
+      transaction.commit();
+    }
+    session.close();
+  }
+}
