--- hw12-WebServer/src/main/java/ru/otus/hw12/hibernate/core/dao/UserDaoException.java
@@ -1,7 +1,0 @@
-package ru.otus.hw12.hibernate.core.dao;
-
-public class UserDaoException extends RuntimeException {
-  public UserDaoException(Exception ex) {
-    super(ex);
-  }
-}
