--- hw12-WebServer/src/main/java/ru/otus/hw12/hibernate/core/sessionmanager/DatabaseSession.java
@@ -1,4 +1,0 @@
-package ru.otus.hw12.hibernate.core.sessionmanager;
-
-public interface DatabaseSession {
-}
