--- hw12-WebServer/src/main/java/ru/otus/hw12/hibernate/core/sessionmanager/SessionManager.java
@@ -1,10 +1,0 @@
-package ru.otus.hw12.hibernate.core.sessionmanager;
-
-public interface SessionManager extends AutoCloseable {
-    void beginSession();
-    void commitSession();
-    void rollbackSession();
-    void close();
-
-    DatabaseSession getCurrentSession();
-}
