--- hw12-WebServer/src/main/java/ru/otus/hw12/hibernate/core/sessionmanager/SessionManagerException.java
@@ -1,12 +1,0 @@
-package ru.otus.hw12.hibernate.core.sessionmanager;
-
-
-public class SessionManagerException extends RuntimeException {
-  public SessionManagerException(String msg) {
-    super(msg);
-  }
-
-  public SessionManagerException(Exception ex) {
-    super(ex);
-  }
-}
