--- hw12-WebServer/src/main/java/ru/otus/hw12/hibernate/hibernate/dao/UserDaoHibernate.java
@@ -1,57 +1,0 @@
-package ru.otus.hw12.hibernate.hibernate.dao;
-
-
-import org.hibernate.Session;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import ru.otus.hw12.hibernate.core.dao.UserDao;
-import ru.otus.hw12.hibernate.core.dao.UserDaoException;
-import ru.otus.hw12.hibernate.core.model.User;
-import ru.otus.hw12.hibernate.core.sessionmanager.SessionManager;
-import ru.otus.hw12.hibernate.hibernate.sessionmanager.DatabaseSessionHibernate;
-import ru.otus.hw12.hibernate.hibernate.sessionmanager.SessionManagerHibernate;
-
-import java.util.Optional;
-
-public class UserDaoHibernate implements UserDao {
-  private static Logger logger = LoggerFactory.getLogger(UserDaoHibernate.class);
-
-  private final SessionManagerHibernate sessionManager;
-
-  public UserDaoHibernate(SessionManagerHibernate sessionManager) {
-    this.sessionManager = sessionManager;
-  }
-
-  @Override
-  public Optional<User> findById(long id) {
-    DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
-    try {
-      return Optional.ofNullable(currentSession.getHibernateSession().find(User.class, id));
-    } catch (Exception e) {
-      logger.error(e.getMessage(), e);
-    }
-    return Optional.empty();
-  }
-
-  @Override
-  public long saveUser(User user) {
-    DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
-    try {
-      Session hibernateSession = currentSession.getHibernateSession();
-      if (user.getId() > 0) {
-        hibernateSession.merge(user);
-      } else {
-        hibernateSession.persist(user);
-      }
-      return user.getId();
-    } catch (Exception e) {
-      logger.error(e.getMessage(), e);
-      throw new UserDaoException(e);
-    }
-  }
-
-  @Override
-  public SessionManager getSessionManager() {
-    return sessionManager;
-  }
-}
