+++ hw12-WebServer/src/main/java/ru/otus/hw12/server/UsersWebServerSimple.java
@@ -56,7 +56,6 @@ public class UsersWebServerSimple implements UsersWebServer {
         handlers.addHandler(resourceHandler);
         handlers.addHandler(applySecurity(servletContextHandler, "/admin", "/api/user/*", "/*"));
 
-
         server.setHandler(handlers);
         return server;
     }
