+++ hw12-WebServer/src/main/java/ru/otus/hw12/servlet/MyAdminServlet.java
@@ -1,7 +1,6 @@
 package ru.otus.hw12.servlet;
 
 import ru.otus.hw12.dao.UserDao;
-import ru.otus.hw12.hibernate.core.model.User;
 import ru.otus.hw12.services.TemplateProcessor;
 
 import javax.servlet.ServletException;
@@ -10,12 +9,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class MyAdminServlet extends HttpServlet {
     private static final String USERS_PAGE_TEMPLATE = "admin.html";
-    private static final String TEMPLATE_ATTR_CREATED_USER = "createdUser";
 
     private final TemplateProcessor templateProcessor;
     private final UserDao userDao;
