--- hw12-WebServer/src/main/java/ru/otus/hw12/servlet/UsersServlet.java
@@ -1,36 +1,0 @@
-package ru.otus.hw12.servlet;
-
-import ru.otus.hw12.dao.UserDaoWebServer;
-import ru.otus.hw12.services.TemplateProcessor;
-
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-
-public class UsersServlet extends HttpServlet {
-
-    private static final String USERS_PAGE_TEMPLATE = "users.html";
-    private static final String TEMPLATE_ATTR_RANDOM_USER = "randomUser";
-
-    private final UserDaoWebServer userDao;
-    private final TemplateProcessor templateProcessor;
-
-    public UsersServlet(TemplateProcessor templateProcessor, UserDaoWebServer userDao) {
-        this.templateProcessor = templateProcessor;
-        this.userDao = userDao;
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse response) throws IOException {
-        Map<String, Object> paramsMap = new HashMap<>();
-      //  userDao.findRandomUser().ifPresent(randomUser -> paramsMap.put(TEMPLATE_ATTR_RANDOM_USER, randomUser));
-
-        response.setContentType("text/html");
-        response.getWriter().println(templateProcessor.getPage(USERS_PAGE_TEMPLATE, paramsMap));
-    }
-
-}
